backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database and ORM
supabase==2.3.4
asyncpg==0.29.0
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OpenAI
openai==1.10.0

# Task Queue
celery==5.3.6
redis==5.0.1

# File handling
python-docx==1.1.0
reportlab==4.0.9
pypdf==4.0.1
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Monitoring
sentry-sdk==1.40.0
